Refactor: move interfaces to utils
