+++ src/utils/interfaces/index.ts
@@ -1,0 +1,21 @@
+export interface IArticle {
+  id: number;
+  title: string;
+  thumbnail: string;
+  description: string;
+  date: string;
+}
+
+export type IArticles = Array<IArticle>;
+
+export interface ApiHandlers {
+  onSuccess(data: IArticles): void;
+  onError(error: Error): void;
+}
+
+export interface FetchRequest {
+  url: string;
+  method?: string;
+  param?: object;
+  handler?: ApiHandlers;
+}
